requirements.txt
streamlit
requests
python-dotenv
rich
pandas
